Fix net.Conn deadlines tearing down the connection when no reads or writes are active
We were unconditionally canceling the read/write contexts when the
deadline timer fired, even if there were no active reads or writes.
We instead only cancel the context if there is an active operation,
otherwise we set a flag so that future calls (without the deadline being
reset) will fail.

Updates tailscale/tailscale#5921