Fix dial panic when ctx is nil
When the ctx is nil, http.NewRequestWithContext returns a "net/http:
nil Context" error and a nil request. In this case, the dial function
panics because it assumes the req is never nil. This checks the
returning error and returns it, so that callers get an error instead
of a panic in that scenario.