Fix unaligned load error on 32-bit architectures
On some 32-bit architectures, 64-bit atomic operations panic when the
value is not aligned properly.

In this package, this causes netConn operations to panic when compiling
with GOARCH=386, since netConn does atomic operations with int64 values
in the netConn struct (namely, with readExpired and writeExpired).

This commit fixes this by moving readExpired and writeExpired to the
beginning of the struct, which makes them properly aligned.