Add `WASM_NO_SANDBOX=on` to disable Chrome sandbox
The CI is currently failing with the following error message:
"No usable sandbox! If you are running on Ubuntu 23.10+ or another Linux
distro that has disabled unprivileged user namespaces..."

This commit updates the wasmbrowsertest to a version that supports the
WASM_NO_SANDBOX environment variable, which disables the Chrome sandbox.
Then it adds the environment variable to `TestWasm` in `conn_test.go`.